Minor bug fix

+++ src/params_all.h
@@ -76,7 +76,7 @@ PARAM(double,qel_s_axial_mass,1030)\
 PARAM(bool,flux_correction,0)\
 PARAM(int,density_model,1)\
 PARAM(int,sf_method,1)\
-PARAM(int,sf_transparency_scale,1)\
+PARAM(double,sf_transparency_scale,1)\
 PARAM(int,sf_transparency_table_idx,1)\
 PARAM(bool,sf_nuclearRecoil,1)\
 PARAM(bool,sf_CoulombDistortion,1)\
